requirements.txt
spotipy==2.24.0
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
flask==3.0.3